requirements.txt
streamlit
tinfoil
python-dotenv